requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Beaker==1.12.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.6
yarl==1.9.2
